Move paymentProcessorAuthorizeNetCreate test function to authorizenet test trait
